don't expose the django secret key to the world
Best writeup of this is probably Dario's email to me:

I thought I'd bring to your attention the fact that the qtile.org repo on
Github is exposing a Django SECRET_KEY in [1]. I don't know if that's the
actual key used, but if it is it may be better to import it from a separate
gitignored module. A bonus would be to also change it to a new string, if
possible, see also [2], [3] and [4].

[1] https://github.com/qtile/qtile.org/blob/master/qtile/settings/common.py#L22
[2] https://docs.djangoproject.com/en/1.10/topics/signing/
[3] https://www.quora.com/What-is-the-SECRET_KEY-in-Django-actually-used-for
[4] https://stackoverflow.com/questions/15170637/effects-of-changing-djangos-secret-key

I've just set up a new variable in heroku, and pushed a new version there.

Reported-by: Dario Giovannetti <dariogiova@gmail.com>
Signed-off-by: Tycho Andersen <tycho@tycho.ws>