dynamically import default_config.py
The default_config.py on our homepage was very old (and even included
dpi_scale, which is not a setting that has ever existed).

This will only rebuild it when the site rebuilds, but I suppose we could
enable a deploy via github actions cron if we really want it updated
regularly.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>